Updated curlcpp (important bug fixes)
This update fixes the underlying issue behind an infinite-stalling bug
that has been around since differencebot was written. The four bots that
download images from ImageNet (difference, advice, grunge, capital) had
the same small logic error whereby the catch block, written to handle the
situation when an image had failed to download, incorrectly took the
exception by value rather than by const reference. While inefficient,
this should not have caused any problems, but curlcpp, which the four
bots all use through libtwitter++'s dependence on it, had a bug in the
copy constructor of the curl_exception object which could cause a
deadlock situation. This went unnoticed for a while, despite the four
bots having serious downtime issues. When this was discovered, the bots
were patched so that the relevant catch block took the exception by
const reference. While this fixed things, it is also prudent to update
curlcpp to the newest version in which this bug has been fixed.

refs JosephP91/curlcpp#100