Fixed unpredictable media upload failure
While uploading media worked in development, as well as on production
when compiled in debug mode, it failed with memory corruption issues on
production in release mode. This turned out to be because curlcpp was
taking lvalue references to temporary std::string objects converted from
string constants, instead of just copying the string pointer.